x86/spec-ctrl: Widen the {xen,last,default}_spec_ctrl fields
Right now, they're all bytes, but MSR_SPEC_CTRL has been steadily gaining new
features.

No functional change.

(cherry picked from commit 45dac88e78e8a2d9d8738eef884fe6730faf9e67)

+++ xen/include/asm-x86/spec_ctrl.h
@@ -94,7 +94,7 @@ extern int8_t opt_l1d_flush;
 extern bool opt_branch_harden;
 
 extern bool bsp_delay_spec_ctrl;
-extern uint8_t default_xen_spec_ctrl;
+extern unsigned int default_xen_spec_ctrl;
 extern uint8_t default_scf;
 
 extern int8_t opt_xpti_hwdom, opt_xpti_domu;
